refactor: move the trace-format surface into dial9::format
